# lib/api-zod/src/generated/types/createEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PromptAnswer } from "./promptAnswer";

export interface CreateEntryBody {
  date: string;
  template?: string;
  answers: PromptAnswer[];
}