# lib/api-zod/src/generated/types/entrySource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EntrySourceBackend } from "./entrySourceBackend";

export interface EntrySource {
  backend: EntrySourceBackend;
  path?: string;
}