# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EntrySource } from "./entrySource";
import type { PromptAnswer } from "./promptAnswer";

export interface JournalEntry {
  id: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  template: string;
  answers: PromptAnswer[];
  /** Full rendered markdown of the entry */
  markdown: string;
  /** ISO timestamp */
  createdAt: string;
  /** ISO timestamp */
  updatedAt: string;
  source: EntrySource;
}