# lib/api-zod/src/generated/types/journalSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JournalSummary {
  totalEntries: number;
  currentStreak: number;
  longestStreak: number;
  /** List of months with entries (YYYY-MM format) */
  activeMonths: string[];
  /** Average presence score (1-10) across all entries */
  averagePresenceScore: number;
}