# lib/api-zod/src/generated/types/storageSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StorageSettingsBackend } from "./storageSettingsBackend";

export interface StorageSettings {
  backend: StorageSettingsBackend;
  /** Path for local file storage */
  localPath?: string;
  /** Google Drive folder ID (future use) */
  googleDriveFolderId?: string;
}