# lib/api-zod/src/generated/types/updateEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PromptAnswer } from "./promptAnswer";

export interface UpdateEntryBody {
  answers: PromptAnswer[];
  markdown?: string;
}